Update README.md
Added the paper "SciTweets - A Dataset and Annotation Framework for Detecting Scientific Online Discourse" (Hafid et al., 2022) which my colleagues and I published at CIKM'22. The contribution includes a dataset which contains scientific claims as they naturally occur on Twitter, as well as an annotation protocol to label scientific claims and a made-publicly-available classifier to detect scientific claims. As the contribution includes both a dataset and a classifier, I included the paper in both the "Claim Detection Models" and "Claim Detection Datasets" sections.